requirements.txt
streamlit>=1.28
pandas
plotly
openpyxl
requests
python-dotenv